General client refactoring
+Network api placed in a separate folder
+Deleted unnecessary file
+Fragments placed in a separate folder
