--- client/app/src/main/java/com/myproj/client/DateSharer.kt
@@ -1,4 +1,0 @@
-package com.myproj.client
-
-class DateSharer {
-}
