+++ client/app/src/main/java/com/myproj/client/fragments/SettingsFragment.kt
@@ -1,10 +1,11 @@
-package com.myproj.client
+package com.myproj.client.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.myproj.client.R
 
 class SettingsFragment : Fragment() {
