+++ client/app/src/main/res/layout/fragment_chart.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:padding="20dp"
     android:paddingTop="40dp"
-    tools:context=".ChartFragment"
+    tools:context=".fragments.ChartFragment"
     android:id="@+id/chart_container">
 
     <LinearLayout
